fix several latent bugs found during code review
- crypto.base64Encode: padding was wrong for inputs whose length mod 3
  was 1 or 2; e.g. base64('A') produced 'QQAA' instead of 'QQ=='. The
  padding check used the post-loop value of `i` against `len`/`len+1`,
  but those equalities never held. Switched to tracking the bytes
  actually consumed in each group. Updated the lib_crypto integration
  test which had been pinned to the buggy output.

- string.format: %s memcpy and %f sprintf wrote into a fixed 4096-byte
  stack buffer with no per-write bounds check; a long argument or a
  large %f could overflow. Switched to snprintf and clamp memcpy to
  remaining space.

- yaml block-seq parser: the "restore L" path was a self-assignment so
  the line view stayed mutated after parsing a virtual mapping.
  Restore from the saved YLine instead.

- vm OP_ADD string concat: u32 length addition could overflow before
  the buffer alloc, leading to a too-small allocation and subsequent
  memcpy overflow. Added an explicit overflow guard.

- array push/insert/rawset: guarded against u32 wrap when length or
  index reaches UINT32_MAX, and made array_ensure_cap's doubling loop
  safe against u32 overflow.

- function.c: cast upvalue_count to size_t before multiplying with
  sizeof(CdoValue) to avoid 32-bit multiplication overflow on large
  upvalue counts.

- Removed two pieces of dead code that were causing -Wunused warnings:
  vm.c's `meta_call_depth` local in OP_CALL and lib/file.c's stale
  `file_methods[]` table that duplicated the real registration list.

Build is now warning-free; all 54 tests pass.

+++ source/object/function.c
@@ -15,7 +15,8 @@ CdoObject *cdo_function_new(u32 param_count, void *bytecode,
     obj->fn.script.bytecode      = bytecode;
     obj->fn.script.upvalue_count = upvalue_count;
     if (upvalue_count > 0 && upvalues) {
-        obj->fn.script.upvalues = cando_alloc(upvalue_count * sizeof(CdoValue));
+        obj->fn.script.upvalues = cando_alloc(
+            (size_t)upvalue_count * sizeof(CdoValue));
         for (u32 i = 0; i < upvalue_count; i++)
             obj->fn.script.upvalues[i] = cdo_value_copy(upvalues[i]);
     }
